+++ frame/system/src/lib.rs
@@ -132,6 +132,8 @@ use codec::{Encode, Decode, FullCodec, EncodeLike};
 use sp_io::TestExternalities;
 
 pub mod offchain;
+#[cfg(test)]
+pub(crate) mod mock;
 
 mod extensions;
 mod weights;
@@ -386,7 +388,7 @@ decl_storage! {
 		ExtrinsicCount: Option<u32>;
 
 		/// The current weight for the block.
-		BlockWeight get(fn block_weight): weight::ExtrinsicsWeight;
+		BlockWeight get(fn block_weight): weights::ExtrinsicsWeight;
 
 		/// Total length (in bytes) for all extrinsics put together, for the current block.
 		AllExtrinsicsLen: Option<u32>;
@@ -1346,115 +1348,12 @@ impl<T: Trait> Lookup for ChainContext<T> {
 }
 
 #[cfg(test)]
-pub(crate) mod tests {
-	use super::*;
-	use sp_std::cell::RefCell;
+mod tests {
+	use crate::*;
+	use mock::{*, Origin};
 	use sp_core::H256;
-	use sp_runtime::{traits::{BlakeTwo256, IdentityLookup, SignedExtension}, testing::Header, DispatchError};
-	use frame_support::{
-		impl_outer_origin, parameter_types, assert_ok, assert_noop,
-		weights::{WithPostDispatchInfo, Pays},
-	};
-
-	impl_outer_origin! {
-		pub enum Origin for Test where system = super {}
-	}
-
-	#[derive(Clone, Eq, PartialEq, Debug)]
-	pub struct Test;
-
-	parameter_types! {
-		pub const BlockHashCount: u64 = 10;
-		pub const MaximumBlockWeight: Weight = 1024;
-		pub const MaximumExtrinsicWeight: Weight = 768;
-		pub const AvailableBlockRatio: Perbill = Perbill::from_percent(75);
-		pub const MaximumBlockLength: u32 = 1024;
-		pub Version: RuntimeVersion = RuntimeVersion {
-			spec_name: sp_version::create_runtime_str!("test"),
-			impl_name: sp_version::create_runtime_str!("system-test"),
-			authoring_version: 1,
-			spec_version: 1,
-			impl_version: 1,
-			apis: sp_version::create_apis_vec!([]),
-			transaction_version: 1,
-		};
-		pub const BlockExecutionWeight: Weight = 10;
-		pub const ExtrinsicBaseWeight: Weight = 5;
-		pub const DbWeight: RuntimeDbWeight = RuntimeDbWeight {
-			read: 10,
-			write: 100,
-		};
-	}
-
-	thread_local!{
-		pub static KILLED: RefCell<Vec<u64>> = RefCell::new(vec![]);
-	}
-
-	pub struct RecordKilled;
-	impl OnKilledAccount<u64> for RecordKilled {
-		fn on_killed_account(who: &u64) { KILLED.with(|r| r.borrow_mut().push(*who)) }
-	}
-
-	#[derive(Debug, codec::Encode, codec::Decode)]
-	pub struct Call;
-
-	impl Dispatchable for Call {
-		type Origin = Origin;
-		type Trait = ();
-		type Info = DispatchInfo;
-		type PostInfo = PostDispatchInfo;
-		fn dispatch(self, _origin: Self::Origin)
-			-> sp_runtime::DispatchResultWithInfo<Self::PostInfo> {
-				panic!("Do not use dummy implementation for dispatch.");
-		}
-	}
-
-	impl Trait for Test {
-		type BaseCallFilter = ();
-		type Origin = Origin;
-		type Call = Call;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = Event<Self>;
-		type BlockHashCount = BlockHashCount;
-		type MaximumBlockWeight = MaximumBlockWeight;
-		type DbWeight = DbWeight;
-		type BlockExecutionWeight = BlockExecutionWeight;
-		type ExtrinsicBaseWeight = ExtrinsicBaseWeight;
-		type MaximumExtrinsicWeight = MaximumExtrinsicWeight;
-		type AvailableBlockRatio = AvailableBlockRatio;
-		type MaximumBlockLength = MaximumBlockLength;
-		type Version = Version;
-		type ModuleToIndex = ();
-		type AccountData = u32;
-		type OnNewAccount = ();
-		type OnKilledAccount = RecordKilled;
-	}
-
-	type System = Module<Test>;
-	type SysEvent = <Test as Trait>::Event;
-
-	const CALL: &<Test as Trait>::Call = &Call;
-
-	fn new_test_ext() -> sp_io::TestExternalities {
-		let mut ext: sp_io::TestExternalities = GenesisConfig::default().build_storage::<Test>().unwrap().into();
-		// Add to each test the initial weight of a block
-		ext.execute_with(|| System::register_extra_weight_unchecked(<Test as Trait>::BlockExecutionWeight::get(), DispatchClass::Mandatory));
-		ext
-	}
-
-	fn normal_weight_limit() -> Weight {
-		<Test as Trait>::AvailableBlockRatio::get() * <Test as Trait>::MaximumBlockWeight::get()
-	}
-
-	fn normal_length_limit() -> u32 {
-		<Test as Trait>::AvailableBlockRatio::get() * <Test as Trait>::MaximumBlockLength::get()
-	}
+	use sp_runtime::DispatchError;
+	use frame_support::weights::WithPostDispatchInfo;
 
 	#[test]
 	fn origin_works() {
@@ -1744,375 +1643,6 @@ pub(crate) mod tests {
 			}
 		})
 	}
-
-	#[test]
-	fn signed_ext_check_nonce_works() {
-		new_test_ext().execute_with(|| {
-			Account::<Test>::insert(1, AccountInfo { nonce: 1, refcount: 0, data: 0 });
-			let info = DispatchInfo::default();
-			let len = 0_usize;
-			// stale
-			assert!(CheckNonce::<Test>(0).validate(&1, CALL, &info, len).is_err());
-			assert!(CheckNonce::<Test>(0).pre_dispatch(&1, CALL, &info, len).is_err());
-			// correct
-			assert!(CheckNonce::<Test>(1).validate(&1, CALL, &info, len).is_ok());
-			assert!(CheckNonce::<Test>(1).pre_dispatch(&1, CALL, &info, len).is_ok());
-			// future
-			assert!(CheckNonce::<Test>(5).validate(&1, CALL, &info, len).is_ok());
-			assert!(CheckNonce::<Test>(5).pre_dispatch(&1, CALL, &info, len).is_err());
-		})
-	}
-
-	#[test]
-	fn signed_ext_check_weight_works_normal_tx() {
-		new_test_ext().execute_with(|| {
-			let normal_limit = normal_weight_limit();
-			let small = DispatchInfo { weight: 100, ..Default::default() };
-			let medium = DispatchInfo {
-				weight: normal_limit - <Test as Trait>::ExtrinsicBaseWeight::get(),
-				..Default::default()
-			};
-			let big = DispatchInfo {
-				weight: normal_limit - <Test as Trait>::ExtrinsicBaseWeight::get() + 1,
-				..Default::default()
-			};
-			let len = 0_usize;
-
-			let reset_check_weight = |i, f, s| {
-				BlockWeight::mutate(|current_weight| {
-					current_weight.put(s, DispatchClass::Normal)
-				});
-				let r = CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, i, len);
-				if f { assert!(r.is_err()) } else { assert!(r.is_ok()) }
-			};
-
-			reset_check_weight(&small, false, 0);
-			reset_check_weight(&medium, false, 0);
-			reset_check_weight(&big, true, 1);
-		})
-	}
-
-	#[test]
-	fn signed_ext_check_weight_refund_works() {
-		new_test_ext().execute_with(|| {
-			// This is half of the max block weight
-			let info = DispatchInfo { weight: 512, ..Default::default() };
-			let post_info = PostDispatchInfo { actual_weight: Some(128), };
-			let len = 0_usize;
-
-			// We allow 75% for normal transaction, so we put 25% - extrinsic base weight
-			BlockWeight::mutate(|current_weight| {
-				current_weight.put(256 - <Test as Trait>::ExtrinsicBaseWeight::get(), DispatchClass::Normal)
-			});
-
-			let pre = CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &info, len).unwrap();
-			assert_eq!(BlockWeight::get().total(), info.weight + 256);
-
-			assert!(
-				CheckWeight::<Test>::post_dispatch(pre, &info, &post_info, len, &Ok(()))
-				.is_ok()
-			);
-			assert_eq!(
-				BlockWeight::get().total(),
-				post_info.actual_weight.unwrap() + 256,
-			);
-		})
-	}
-
-	#[test]
-	fn signed_ext_check_weight_actual_weight_higher_than_max_is_capped() {
-		new_test_ext().execute_with(|| {
-			let info = DispatchInfo { weight: 512, ..Default::default() };
-			let post_info = PostDispatchInfo { actual_weight: Some(700), };
-			let len = 0_usize;
-
-			BlockWeight::mutate(|current_weight| {
-				current_weight.put(128, DispatchClass::Normal)
-			});
-
-			let pre = CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &info, len).unwrap();
-			assert_eq!(
-				BlockWeight::get().total(),
-				info.weight + 128 + <Test as Trait>::ExtrinsicBaseWeight::get(),
-			);
-
-			assert!(
-				CheckWeight::<Test>::post_dispatch(pre, &info, &post_info, len, &Ok(()))
-				.is_ok()
-			);
-			assert_eq!(
-				BlockWeight::get().total(),
-				info.weight + 128 + <Test as Trait>::ExtrinsicBaseWeight::get(),
-			);
-		})
-	}
-
-	#[test]
-	fn zero_weight_extrinsic_still_has_base_weight() {
-		new_test_ext().execute_with(|| {
-			let free = DispatchInfo { weight: 0, ..Default::default() };
-			let len = 0_usize;
-
-			// Initial weight from `BlockExecutionWeight`
-			assert_eq!(System::block_weight().total(), <Test as Trait>::BlockExecutionWeight::get());
-			let r = CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &free, len);
-			assert!(r.is_ok());
-			assert_eq!(
-				System::block_weight().total(),
-				<Test as Trait>::ExtrinsicBaseWeight::get() + <Test as Trait>::BlockExecutionWeight::get()
-			);
-		})
-	}
-
-	#[test]
-	fn mandatory_extrinsic_doesnt_care_about_limits() {
-		fn check(call: impl FnOnce(&DispatchInfo, usize)) {
-			new_test_ext().execute_with(|| {
-				let max = DispatchInfo {
-					weight: Weight::max_value(),
-					class: DispatchClass::Mandatory,
-					..Default::default()
-				};
-				let len = 0_usize;
-
-				call(&max, len);
-			});
-		}
-
-		check(|max, len| {
-			assert_ok!(CheckWeight::<Test>::do_pre_dispatch(max, len));
-			assert_eq!(System::block_weight().total(), Weight::max_value());
-			assert!(System::block_weight().total() > <Test as Trait>::MaximumBlockWeight::get());
-		});
-		check(|max, len| {
-			assert_ok!(CheckWeight::<Test>::do_validate(max, len));
-		});
-	}
-
-	#[test]
-	fn normal_extrinsic_limited_by_maximum_extrinsic_weight() {
-		new_test_ext().execute_with(|| {
-			let max = DispatchInfo {
-				weight: MaximumExtrinsicWeight::get() + 1,
-				class: DispatchClass::Normal,
-				..Default::default()
-			};
-			let len = 0_usize;
-
-			assert_noop!(
-				CheckWeight::<Test>::do_validate(&max, len),
-				InvalidTransaction::ExhaustsResources
-			);
-		});
-	}
-
-	#[test]
-	fn operational_extrinsic_limited_by_operational_space_limit() {
-		new_test_ext().execute_with(|| {
-			let operational_limit = CheckWeight::<Test>::get_dispatch_limit_ratio(
-				DispatchClass::Operational
-			) * <Test as Trait>::MaximumBlockWeight::get();
-			let base_weight = <Test as Trait>::ExtrinsicBaseWeight::get();
-			let block_base = <Test as Trait>::BlockExecutionWeight::get();
-
-			let weight = operational_limit - base_weight - block_base;
-			let okay = DispatchInfo {
-				weight,
-				class: DispatchClass::Operational,
-				..Default::default()
-			};
-			let max = DispatchInfo {
-				weight: weight + 1,
-				class: DispatchClass::Operational,
-				..Default::default()
-			};
-			let len = 0_usize;
-
-			assert_eq!(
-				CheckWeight::<Test>::do_validate(&okay, len),
-				Ok(ValidTransaction {
-					priority: CheckWeight::<Test>::get_priority(&okay),
-					..Default::default()
-				})
-			);
-			assert_noop!(
-				CheckWeight::<Test>::do_validate(&max, len),
-				InvalidTransaction::ExhaustsResources
-			);
-		});
-	}
-
-	#[test]
-	fn register_extra_weight_unchecked_doesnt_care_about_limits() {
-		new_test_ext().execute_with(|| {
-			System::register_extra_weight_unchecked(Weight::max_value(), DispatchClass::Normal);
-			assert_eq!(System::block_weight().total(), Weight::max_value());
-			assert!(System::block_weight().total() > <Test as Trait>::MaximumBlockWeight::get());
-		});
-	}
-
-	#[test]
-	fn full_block_with_normal_and_operational() {
-		new_test_ext().execute_with(|| {
-			// Max block is 1024
-			// Max normal is 768 (75%)
-			// 10 is taken for block execution weight
-			// So normal extrinsic can be 758 weight (-5 for base extrinsic weight)
-			// And Operational can be 256 to produce a full block (-5 for base)
-			let max_normal = DispatchInfo { weight: 753, ..Default::default() };
-			let rest_operational = DispatchInfo { weight: 251, class: DispatchClass::Operational, ..Default::default() };
-
-			let len = 0_usize;
-
-			assert_ok!(CheckWeight::<Test>::do_pre_dispatch(&max_normal, len));
-			assert_eq!(System::block_weight().total(), 768);
-			assert_ok!(CheckWeight::<Test>::do_pre_dispatch(&rest_operational, len));
-			assert_eq!(<Test as Trait>::MaximumBlockWeight::get(), 1024);
-			assert_eq!(System::block_weight().total(), <Test as Trait>::MaximumBlockWeight::get());
-			// Checking single extrinsic should not take current block weight into account.
-			assert_eq!(CheckWeight::<Test>::check_extrinsic_weight(&rest_operational), Ok(()));
-		});
-	}
-
-	#[test]
-	fn dispatch_order_does_not_effect_weight_logic() {
-		new_test_ext().execute_with(|| {
-			// We switch the order of `full_block_with_normal_and_operational`
-			let max_normal = DispatchInfo { weight: 753, ..Default::default() };
-			let rest_operational = DispatchInfo { weight: 251, class: DispatchClass::Operational, ..Default::default() };
-
-			let len = 0_usize;
-
-			assert_ok!(CheckWeight::<Test>::do_pre_dispatch(&rest_operational, len));
-			// Extra 15 here from block execution + base extrinsic weight
-			assert_eq!(System::block_weight().total(), 266);
-			assert_ok!(CheckWeight::<Test>::do_pre_dispatch(&max_normal, len));
-			assert_eq!(<Test as Trait>::MaximumBlockWeight::get(), 1024);
-			assert_eq!(System::block_weight().total(), <Test as Trait>::MaximumBlockWeight::get());
-		});
-	}
-
-	#[test]
-	fn operational_works_on_full_block() {
-		new_test_ext().execute_with(|| {
-			// An on_initialize takes up the whole block! (Every time!)
-			System::register_extra_weight_unchecked(Weight::max_value(), DispatchClass::Mandatory);
-			let dispatch_normal = DispatchInfo { weight: 251, class: DispatchClass::Normal, ..Default::default() };
-			let dispatch_operational = DispatchInfo { weight: 251, class: DispatchClass::Operational, ..Default::default() };
-			let len = 0_usize;
-
-			assert_noop!(CheckWeight::<Test>::do_pre_dispatch(&dispatch_normal, len), InvalidTransaction::ExhaustsResources);
-			// Thank goodness we can still do an operational transaction to possibly save the blockchain.
-			assert_ok!(CheckWeight::<Test>::do_pre_dispatch(&dispatch_operational, len));
-			// Not too much though
-			assert_noop!(CheckWeight::<Test>::do_pre_dispatch(&dispatch_operational, len), InvalidTransaction::ExhaustsResources);
-			// Even with full block, validity of single transaction should be correct.
-			assert_eq!(CheckWeight::<Test>::check_extrinsic_weight(&dispatch_operational), Ok(()));
-		});
-	}
-
-	#[test]
-	fn signed_ext_check_weight_works_operational_tx() {
-		new_test_ext().execute_with(|| {
-			let normal = DispatchInfo { weight: 100, ..Default::default() };
-			let op = DispatchInfo { weight: 100, class: DispatchClass::Operational, pays_fee: Pays::Yes };
-			let len = 0_usize;
-			let normal_limit = normal_weight_limit();
-
-			// given almost full block
-			BlockWeight::mutate(|current_weight| {
-				current_weight.put(normal_limit, DispatchClass::Normal)
-			});
-			// will not fit.
-			assert!(CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &normal, len).is_err());
-			// will fit.
-			assert!(CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &op, len).is_ok());
-
-			// likewise for length limit.
-			let len = 100_usize;
-			AllExtrinsicsLen::put(normal_length_limit());
-			assert!(CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &normal, len).is_err());
-			assert!(CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, &op, len).is_ok());
-		})
-	}
-
-	#[test]
-	fn signed_ext() {
-		new_test_ext().execute_with(|| {
-			let normal = DispatchInfo { weight: 100, class: DispatchClass::Normal, pays_fee: Pays::Yes };
-			let op = DispatchInfo { weight: 100, class: DispatchClass::Operational, pays_fee: Pays::Yes };
-			let len = 0_usize;
-
-			let priority = CheckWeight::<Test>(PhantomData)
-				.validate(&1, CALL, &normal, len)
-				.unwrap()
-				.priority;
-			assert_eq!(priority, 100);
-
-			let priority = CheckWeight::<Test>(PhantomData)
-				.validate(&1, CALL, &op, len)
-				.unwrap()
-				.priority;
-			assert_eq!(priority, u64::max_value() / 2);
-		})
-	}
-
-	#[test]
-	fn signed_ext_check_weight_block_size_works() {
-		new_test_ext().execute_with(|| {
-			let normal = DispatchInfo::default();
-			let normal_limit = normal_weight_limit() as usize;
-			let reset_check_weight = |tx, s, f| {
-				AllExtrinsicsLen::put(0);
-				let r = CheckWeight::<Test>(PhantomData).pre_dispatch(&1, CALL, tx, s);
-				if f { assert!(r.is_err()) } else { assert!(r.is_ok()) }
-			};
-
-			reset_check_weight(&normal, normal_limit - 1, false);
-			reset_check_weight(&normal, normal_limit, false);
-			reset_check_weight(&normal, normal_limit + 1, true);
-
-			// Operational ones don't have this limit.
-			let op = DispatchInfo { weight: 0, class: DispatchClass::Operational, pays_fee: Pays::Yes };
-			reset_check_weight(&op, normal_limit, false);
-			reset_check_weight(&op, normal_limit + 100, false);
-			reset_check_weight(&op, 1024, false);
-			reset_check_weight(&op, 1025, true);
-		})
-	}
-
-	#[test]
-	fn signed_ext_check_era_should_work() {
-		new_test_ext().execute_with(|| {
-			// future
-			assert_eq!(
-				CheckEra::<Test>::from(Era::mortal(4, 2)).additional_signed().err().unwrap(),
-				InvalidTransaction::AncientBirthBlock.into(),
-			);
-
-			// correct
-			System::set_block_number(13);
-			<BlockHash<Test>>::insert(12, H256::repeat_byte(1));
-			assert!(CheckEra::<Test>::from(Era::mortal(4, 12)).additional_signed().is_ok());
-		})
-	}
-
-	#[test]
-	fn signed_ext_check_era_should_change_longevity() {
-		new_test_ext().execute_with(|| {
-			let normal = DispatchInfo { weight: 100, class: DispatchClass::Normal, pays_fee: Pays::Yes };
-			let len = 0_usize;
-			let ext = (
-				CheckWeight::<Test>(PhantomData),
-				CheckEra::<Test>::from(Era::mortal(16, 256)),
-			);
-			System::set_block_number(17);
-			<BlockHash<Test>>::insert(16, H256::repeat_byte(1));
-
-			assert_eq!(ext.validate(&1, CALL, &normal, len).unwrap().longevity, 15);
-		})
-	}
-
 
 	#[test]
 	fn set_code_checks_works() {
